webapp/flask-backend/requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.1.7
Flask==2.2.3
Flask-Cors==3.0.10
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
six==1.16.0
sqlparse==0.4.3
Werkzeug==2.2.3
